Pipe generation, Fish Rotation
North and South pipes now draw on the canvas. have set gap spacing and spacing between each pipe. Where the gap is on the y axis though is randomized but the gap will always be at a height of 80 pixels.

Fish also rotates after x amount of time not tapping.